Refactored correctors and added GP corrector
