+++ lightkurve/correctors/__init__.py
@@ -1,0 +1,4 @@
+from .pldcorrector import *
+from .sffcorrector import *
+from .cbvcorrector import *
+from .gpcorrector import *
