+++ lightkurve/lightcurve.py
@@ -1214,17 +1214,21 @@ class LightCurve(object):
             Instance of a Corrector class, which typically provides `correct()`
             and `diagnose()` methods.
         """
-        allowed_methods = ["sff"]
-        if method == "pld":
+        allowed_methods = ["sff", "gp"]
+        if method.lower() == "pld":
             raise ValueError("The 'pld' method can only be used on "
                              "`TargetPixelFile` objects, not `LightCurve` objects.")
-        if method not in allowed_methods:
+        if method.lower() not in allowed_methods:
             raise ValueError(("Unrecognized method '{0}'\n"
                               "allowed methods are: {1}")
                              .format(method, allowed_methods))
-        if method == "sff":
+        if method.lower() == "sff":
             from .correctors import SFFCorrector
             return SFFCorrector(self)
+
+        if method.lower() == "gp":
+            from .correctors import GPCorrector
+            return GPCorrector(self)
 
 
 class FoldedLightCurve(LightCurve):
